slack/requirements.txt
slackclient
slackeventsapi
python-dotenv
flask